+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/ImplementationBridgeHelpers.java
@@ -4,9 +4,11 @@
 package com.azure.cosmos.implementation;
 
 import com.azure.cosmos.BulkProcessingOptions;
+import com.azure.cosmos.BulkProcessingThresholds;
 import com.azure.cosmos.CosmosAsyncClient;
 import com.azure.cosmos.CosmosClient;
 import com.azure.cosmos.CosmosClientBuilder;
+import com.azure.cosmos.implementation.batch.PartitionScopeThresholds;
 import com.azure.cosmos.implementation.routing.PartitionKeyInternal;
 import com.azure.cosmos.implementation.spark.OperationContextAndListenerTuple;
 import com.azure.cosmos.models.CosmosItemRequestOptions;
@@ -15,6 +17,8 @@ import com.azure.cosmos.models.CosmosQueryRequestOptions;
 import com.azure.cosmos.models.PartitionKey;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.util.concurrent.ConcurrentMap;
 
 public class ImplementationBridgeHelpers {
     private static Logger logger = LoggerFactory.getLogger(ImplementationBridgeHelpers.class);
@@ -28,7 +32,7 @@ public class ImplementationBridgeHelpers {
 
         public static void setCosmosClientBuilderAccessor(final CosmosClientBuilderAccessor newAccessor) {
             if (accessor != null) {
-                throw new IllegalStateException("CosmosClientBuilderHelper accessor already initialized!");
+                throw new IllegalStateException("CosmosClientBuilder accessor already initialized!");
             }
 
             accessor = newAccessor;
@@ -36,7 +40,7 @@ public class ImplementationBridgeHelpers {
 
         static CosmosClientBuilderAccessor getCosmosClientBuilderAccessor() {
             if (accessor == null) {
-                throw new IllegalStateException("CosmosClientBuilderHelper accessor is not initialized yet!");
+                throw new IllegalStateException("CosmosClientBuilder accessor is not initialized yet!");
             }
 
             return accessor;
@@ -60,7 +64,7 @@ public class ImplementationBridgeHelpers {
 
         public static void setPartitionKeyAccessor(final PartitionKeyAccessor newAccessor) {
             if (accessor != null) {
-                throw new IllegalStateException("PartitionKeyHelper accessor already initialized!");
+                throw new IllegalStateException("PartitionKey accessor already initialized!");
             }
 
             accessor = newAccessor;
@@ -68,7 +72,7 @@ public class ImplementationBridgeHelpers {
 
         public static PartitionKeyAccessor getPartitionKeyAccessor() {
             if (accessor == null) {
-                throw new IllegalStateException("PartitionKeyHelper accessor is not initialized!");
+                throw new IllegalStateException("PartitionKey accessor is not initialized!");
             }
 
             return accessor;
@@ -84,12 +88,12 @@ public class ImplementationBridgeHelpers {
 
         private CosmosQueryRequestOptionsHelper() {}
         static {
-            ensureClassLoaded(CosmosQueryRequestOptionsHelper.class);
+            ensureClassLoaded(CosmosQueryRequestOptions.class);
         }
 
         public static void setCosmosQueryRequestOptionsAccessor(final CosmosQueryRequestOptionsAccessor newAccessor) {
             if (accessor != null) {
-                throw new IllegalStateException("CosmosQueryRequestOptionsHelper accessor already initialized!");
+                throw new IllegalStateException("CosmosQueryRequestOptions accessor already initialized!");
             }
 
             accessor = newAccessor;
@@ -97,7 +101,7 @@ public class ImplementationBridgeHelpers {
 
         public static CosmosQueryRequestOptionsAccessor getCosmosQueryRequestOptionsAccessor() {
             if (accessor == null) {
-                throw new IllegalStateException("CosmosQueryRequestOptionsHelper accessor is not initialized yet!");
+                throw new IllegalStateException("CosmosQueryRequestOptions accessor is not initialized yet!");
             }
 
             return accessor;
@@ -114,12 +118,12 @@ public class ImplementationBridgeHelpers {
 
         private CosmosItemRequestOptionsHelper() {}
         static {
-            ensureClassLoaded(CosmosQueryRequestOptionsHelper.class);
+            ensureClassLoaded(CosmosItemRequestOptions.class);
         }
 
         public static void setCosmosItemRequestOptionsAccessor(final CosmosItemRequestOptionsAccessor newAccessor) {
             if (accessor != null) {
-                throw new IllegalStateException("CosmosQueryRequestOptionsHelper accessor already initialized!");
+                throw new IllegalStateException("CosmosItemRequestOptions accessor already initialized!");
             }
 
             accessor = newAccessor;
@@ -127,7 +131,7 @@ public class ImplementationBridgeHelpers {
 
         public static CosmosItemRequestOptionsAccessor getCosmosItemRequestOptionsAccessor() {
             if (accessor == null) {
-                throw new IllegalStateException("CosmosQueryRequestOptionsHelper accessor is not initialized yet!");
+                throw new IllegalStateException("CosmosItemRequestOptions accessor is not initialized yet!");
             }
 
             return accessor;
@@ -145,12 +149,12 @@ public class ImplementationBridgeHelpers {
 
         private CosmosBulkProcessingOptionsHelper() {}
         static {
-            ensureClassLoaded(CosmosQueryRequestOptionsHelper.class);
+            ensureClassLoaded(BulkProcessingOptions.class);
         }
 
         public static void setCosmosBulkProcessingOptionAccessor(final CosmosBulkProcessingOptionAccessor newAccessor) {
             if (accessor != null) {
-                throw new IllegalStateException("CosmosQueryRequestOptionsHelper accessor already initialized!");
+                throw new IllegalStateException("BulkProcessingOptions accessor already initialized!");
             }
 
             accessor = newAccessor;
@@ -158,7 +162,7 @@ public class ImplementationBridgeHelpers {
 
         public static CosmosBulkProcessingOptionAccessor getCosmosBulkProcessingOptionAccessor() {
             if (accessor == null) {
-                throw new IllegalStateException("CosmosQueryRequestOptionsHelper accessor is not initialized yet!");
+                throw new IllegalStateException("BulkProcessingOptions accessor is not initialized yet!");
             }
 
             return accessor;
@@ -182,7 +186,7 @@ public class ImplementationBridgeHelpers {
 
         public static void setCosmosItemResponseBuilderAccessor(final CosmosItemResponseBuilderAccessor newAccessor) {
             if (accessor != null) {
-                throw new IllegalStateException("CosmosItemResponseBuilder accessor already initialized!");
+                throw new IllegalStateException("CosmosItemResponse accessor already initialized!");
             }
 
             accessor = newAccessor;
@@ -190,7 +194,7 @@ public class ImplementationBridgeHelpers {
 
         public static CosmosItemResponseBuilderAccessor getCosmosItemResponseBuilderAccessor() {
             if (accessor == null) {
-                throw new IllegalStateException("CosmosItemResponseBuilder accessor is not initialized yet!");
+                throw new IllegalStateException("CosmosItemResponse accessor is not initialized yet!");
             }
 
             return accessor;
@@ -221,7 +225,7 @@ public class ImplementationBridgeHelpers {
 
         public static void setCosmosClientAccessor(final CosmosClientAccessor newAccessor) {
             if (accessor != null) {
-                throw new IllegalStateException("CosmosClientAccessor accessor already initialized!");
+                throw new IllegalStateException("CosmosClient accessor already initialized!");
             }
 
             accessor = newAccessor;
@@ -229,7 +233,7 @@ public class ImplementationBridgeHelpers {
 
         public static CosmosClientAccessor geCosmosClientAccessor() {
             if (accessor == null) {
-                throw new IllegalStateException("CosmosClientAccessor accessor is not initialized yet!");
+                throw new IllegalStateException("CosmosClient accessor is not initialized yet!");
             }
 
             return accessor;
@@ -237,6 +241,36 @@ public class ImplementationBridgeHelpers {
 
         public interface CosmosClientAccessor {
             CosmosAsyncClient getCosmosAsyncClient(CosmosClient cosmosClient);
+        }
+    }
+
+    public static final class BulkProcessingThresholdsHelper {
+        private static BulkProcessingThresholdsAccessor accessor;
+
+        private BulkProcessingThresholdsHelper() {}
+        static {
+            ensureClassLoaded(BulkProcessingThresholds.class);
+        }
+
+        public static void setBulkProcessingThresholdsAccessor(final BulkProcessingThresholdsAccessor newAccessor) {
+            if (accessor != null) {
+                throw new IllegalStateException("BulkProcessingThresholds accessor already initialized!");
+            }
+
+            accessor = newAccessor;
+        }
+
+        public static BulkProcessingThresholdsAccessor getBulkProcessingThresholdsAccessor() {
+            if (accessor == null) {
+                throw new IllegalStateException("BulkProcessingThresholds accessor is not initialized yet!");
+            }
+
+            return accessor;
+        }
+
+        public interface BulkProcessingThresholdsAccessor {
+            <T> ConcurrentMap<String, PartitionScopeThresholds<T>> getPartitionScopeThresholds(
+                BulkProcessingThresholds<T> thresholds);
         }
     }