+++ sdk/textanalytics/azure-ai-textanalytics/src/main/java/com/azure/ai/textanalytics/models/ExtractKeyPhrasesActionResult.java
@@ -6,16 +6,18 @@ package com.azure.ai.textanalytics.models;
 
 import com.azure.ai.textanalytics.implementation.ExtractKeyPhrasesActionResultPropertiesHelper;
 import com.azure.ai.textanalytics.util.ExtractKeyPhrasesResultCollection;
+import com.azure.core.annotation.Immutable;
 
 /**
  * The {@link ExtractKeyPhrasesActionResult} model.
  */
+@Immutable
 public final class ExtractKeyPhrasesActionResult extends TextAnalyticsActionResult {
-    private ExtractKeyPhrasesResultCollection documentResults;
+    private ExtractKeyPhrasesResultCollection documentsResults;
 
     static {
         ExtractKeyPhrasesActionResultPropertiesHelper.setAccessor(
-            (actionsResult, documentResults) -> actionsResult.setDocumentResults(documentResults));
+            (actionResult, documentsResults) -> actionResult.setDocumentsResults(documentsResults));
     }
 
     /**
@@ -26,12 +28,12 @@ public final class ExtractKeyPhrasesActionResult extends TextAnalyticsActionResu
      * @throws TextAnalyticsException if result has {@code isError} equals to true and when a non-error property
      * was accessed.
      */
-    public ExtractKeyPhrasesResultCollection getDocumentResults() {
+    public ExtractKeyPhrasesResultCollection getDocumentsResults() {
         throwExceptionIfError();
-        return documentResults;
+        return documentsResults;
     }
 
-    private void setDocumentResults(ExtractKeyPhrasesResultCollection documentResults) {
-        this.documentResults = documentResults;
+    private void setDocumentsResults(ExtractKeyPhrasesResultCollection documentsResults) {
+        this.documentsResults = documentsResults;
     }
 }