+++ sdk/core/azure-core-serializer-avro-apache/CHANGELOG.md
@@ -1,7 +1,13 @@
 # Release History
 
-## 1.0.0-beta.12 (Unreleased)
+## 1.0.0-beta.13 (Unreleased)
 
+
+## 1.0.0-beta.12 (2021-07-01)
+
+### Dependency Updates
+
+- Upgraded `azure-core` from `1.17.0` to `1.18.0`.
 
 ## 1.0.0-beta.11 (2021-06-07)